Move widgets to appropriate package
