+++ app/src/main/res/layout/sheet_custom_prescription_entry.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<org.simple.clinic.bp.entry.LinearLayoutWithPreImeKeyEventListener android:id="@+id/customprescription_root"
+<org.simple.clinic.widgets.LinearLayoutWithPreImeKeyEventListener android:id="@+id/customprescription_root"
   xmlns:android="http://schemas.android.com/apk/res/android"
   xmlns:tools="http://schemas.android.com/tools"
   android:layout_width="match_parent"
@@ -106,4 +106,4 @@
     android:text="@string/customprescription_save_drug"
     android:textColor="@color/white"
     tools:ignore="UnusedAttribute" />
-</org.simple.clinic.bp.entry.LinearLayoutWithPreImeKeyEventListener>
+</org.simple.clinic.widgets.LinearLayoutWithPreImeKeyEventListener>
